Player projectile fixed. The slope that determines its movement angle was incrementing x and y which were ints, not doubles. This reduced precision by a LOT, because the slope could only have whole numbers. Changing position vars to doubles allowed accurate precision so that the projectile could always go to the mouse click pos, no matter what.

+++ src/util/MyMath.java
@@ -107,7 +107,7 @@ public class MyMath {
     }
 
 
-    public static double getAngle(int x1, int y1, int x2, int y2) {
+    public static double getAngle(double x1, double y1, double x2, double y2) {
         double angle = betterRound(Math.toDegrees(Math.atan2(y2 - y1, x2 - x1)));
         angle = angleWrap(angle);
